Reorganization and refinements
