+++ src/shared/components/NameInput.tsx
