+++ src/shared/components/ShowTypeIcon.tsx
