+++ src/shared/containers/header/Header.css
@@ -1,5 +1,7 @@
-.app-header {
-  background-color: #FDD835;
+.appHeaderBar {
+  /*background-color: #FDD835;*/
+  background-color: #f06543;
+  background-image: linear-gradient(315deg, #f06543 0%, #ffbe3d 74%);
   color: #fff;
 }
